+++ libs/community/langchain_community/memory/zep_memory.py
@@ -2,8 +2,9 @@ from __future__ import annotations
 
 from typing import Any, Dict, Optional
 
-from langchain.memory import ConversationBufferMemory
-from langchain.memory.chat_message_histories import ZepChatMessageHistory
+from langchain_core.memory.buffer import ConversationBufferMemory
+
+from langchain_community.chat_message_histories.zep import ZepChatMessageHistory
 
 
 class ZepMemory(ConversationBufferMemory):
