+++ libs/core/langchain_core/memory/__init__.py
@@ -9,5 +9,3 @@ __all__ = [
     "ChatMessageHistory",
     "get_prompt_input_key",
 ]
-
-
